Update README.md
Fixed the typo.

The line we drew for inclusion of terms is whether the terms are useful for describing things in the Nulecule context - even if "aggregation" is not defined in the spec or "below the spec" - it's still a useful term to know. My only request is that we define it correctly.